src/main/java/com/kaya/BookStore/service/BookService.java
package com.kaya.BookStore.service;

import com.kaya.BookStore.dto.BookDto;
import com.kaya.BookStore.repository.BookRepository;
import org.modelmapper.ModelMapper;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;

@Service
public class BookService {

    private BookRepository bookRepository;

    private ModelMapper modelMapper;

    public BookService(BookRepository bookRepository, ModelMapper modelMapper) {
        this.bookRepository = bookRepository;
        this.modelMapper = modelMapper;
    }

    public List<BookDto> getBooks() {
        return new ArrayList<>();
    }
}
